xml/parsing/manakai/condsects-4.dat

#data
<!DOCTYPE x []]]
#errors
1;1;no XML decl;;;s
1;14;string in internal subset;;;m
1;17;parser:EOF;;;m
1;17;parser:EOF;;;m
1;17;no root element;;;m
#document
| <!DOCTYPE x>
